+++ Homework_3/src/Zoo.java
@@ -1,4 +1,6 @@
 import animals.*;
+import animals.aviarySizeEnum.AviarySize;
+import animals.exception.WrongFoodException;
 import food.*;
 
 import java.util.Map;
