+++ Homework_3/src/animals/Aviary.java
@@ -1,5 +1,6 @@
 package animals;
 
+import animals.aviarySizeEnum.AviarySize;
 import java.util.HashMap;
 import java.util.Map;
