+++ Homework_3/src/animals/Carnivore.java
@@ -1,5 +1,7 @@
 package animals;
 
+import animals.aviarySizeEnum.AviarySize;
+
 public abstract class Carnivore extends Animals {
 
     public Carnivore(String uniqueIDName, AviarySize aviarySize, int hungerLevel, int thirst) {
