+++ Homework_3/src/animals/Herbivore.java
@@ -1,5 +1,7 @@
 package animals;
 
+import animals.aviarySizeEnum.AviarySize;
+
 public abstract class Herbivore extends Animals {
 
     public Herbivore(String uniqueIDName, AviarySize aviarySize, int hungerLevel, int thirst) {
