+++ Homework_3/src/animals/aviarySizeEnum/AviarySize.java
@@ -1,4 +1,4 @@
-package animals;
+package animals.aviarySizeEnum;
 
 public enum AviarySize {
