+++ Homework_3/src/animals/exception/WrongFoodException.java
@@ -1,4 +1,4 @@
-package animals;
+package animals.exception;
 
 public class WrongFoodException extends Exception {
